requirements.txt
google-adk 
google-genai   
python-dotenv